# gzip: escape heap allocation 