Hello,

I've recently taken an interest in both this project and Prometheus, and I've been following PRs on both repositories. 
`parseEncodings` is used to parse the 'Accept-Encoding' from the header of an HTTP request.
I was inspired by a recent change merged into Prometheus (see: prometheus/prometheus#16001) that demonstrates the benefits of this approach, and I propose a similar update for this project. seems this change could be beneficial to Pyroscope too.

I look forward to your feedback and any suggestions you might have. Thank you : )

---------
```
go test -benchmem -run=^$ -count 10 -bench '^BenchmarkParseEncodings$' ./... > new
git checkout 6e31e1b27dbbc1f9f324ae3a386c08adee45811d
go test -benchmem -run=^$ -count 10 -bench '^BenchmarkParseEncodings$' ./... > old
benchstat old new
```

```shell
goos: darwin
goarch: arm64
pkg: github.com/grafana/pyroscope/pkg/util/gziphandler
                  │     old     │                new                 │
                  │   sec/op    │   sec/op     vs base               │
ParseEncodings-10   145.1m ± 2%   136.3m ± 1%  -6.02% (p=0.000 n=10)

                  │     old      │                 new                 │
                  │     B/op     │     B/op      vs base               │
ParseEncodings-10   222.0Mi ± 0%   206.0Mi ± 0%  -7.21% (p=0.000 n=10)

                  │     old     │                new                 │
                  │  allocs/op  │  allocs/op   vs base               │
ParseEncodings-10   3.146M ± 0%   3.146M ± 0%  +0.00% (p=0.000 n=10)
```